refactor: 中央化擴充功能相關的 API 和錯誤碼
- 將 API 和錯誤碼相關的常數從舊的 api.js 檔案中移除，並分別移至新的 extension 目錄下的檔案中，以提高模組化和可維護性。
- 新增 accountApi.js、authMode.js、driveSyncErrorCodes.js 和 notionApi.js 檔案，這些檔案專門用於擴充功能的背景處理和頁面共享，避免 Content Scripts 引入不必要的模組。
- 更新相關的導入路徑，確保所有使用這些常數的檔案都指向新的位置，並刪除不再使用的 api.js 檔案。
