+++ scripts/auth/auth.js
@@ -19,7 +19,7 @@
 /* global chrome */
 
 import { BUILD_ENV } from '../config/env.js';
-import { ACCOUNT_API } from '../config/api.js';
+import { ACCOUNT_API } from '../config/extension/accountApi.js';
 import { RUNTIME_ACTIONS } from '../config/runtimeActions.js';
 import { setAccountSession, setAccountProfile, clearAccountSession } from './accountSession.js';
 import { showError, showLoading, showSuccess } from './callbackStatusView.js';
