+++ scripts/config/extension/authMode.js
@@ -1,0 +1,10 @@
+/**
+ * 認證模式枚舉。
+ *
+ * 供 Background 與 extension pages 共用，
+ * Content Scripts MUST NOT import 此模組。
+ */
+export const AuthMode = {
+  OAUTH: 'oauth',
+  MANUAL: 'manual',
+};
