+++ scripts/utils/notionAuth.js
@@ -1,6 +1,7 @@
 /* global chrome */
 import Logger from './Logger.js';
-import { AuthMode, NOTION_OAUTH } from '../config/api.js';
+import { AuthMode } from '../config/extension/authMode.js';
+import { NOTION_OAUTH } from '../config/extension/notionAuth.js';
 import { BUILD_ENV } from '../config/env.js';
 import { RUNTIME_ACTIONS } from '../config/runtimeActions.js';
 import { sanitizeApiError } from './securityUtils.js';
